+++ src/eps/tutorials/ex49.c
@@ -1,0 +1,171 @@
+/*
+   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+   SLEPc - Scalable Library for Eigenvalue Problem Computations
+   Copyright (c) 2002-2021, Universitat Politecnica de Valencia, Spain
+
+   This file is part of SLEPc.
+   SLEPc is distributed under a 2-clause BSD license (see LICENSE).
+   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+*/
+
+static char help[] = "User-defined split preconditioner when solving a generalized eigenproblem.\n\n"
+  "The command line options are:\n"
+  "  -n <n>, where <n> = number of grid subdivisions in x dimension.\n"
+  "  -m <m>, where <m> = number of grid subdivisions in y dimension.\n\n";
+
+#include <slepceps.h>
+
+int main(int argc,char **argv)
+{
+  Mat            A,B,A0,B0,mats[2]; /* problem matrices and sparser approximations */
+  EPS            eps;               /* eigenproblem solver context */
+  ST             st;
+  PetscInt       N,n=24,m,Istart,Iend,II,i,j;
+  PetscBool      flag,terse;
+  PetscErrorCode ierr;
+
+  ierr = SlepcInitialize(&argc,&argv,(char*)0,help);if (ierr) return ierr;
+
+  ierr = PetscOptionsGetInt(NULL,NULL,"-n",&n,NULL);CHKERRQ(ierr);
+  ierr = PetscOptionsGetInt(NULL,NULL,"-m",&m,&flag);CHKERRQ(ierr);
+  if (!flag) m=n;
+  N = n*m;
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"\nGHEP with split preconditioner, N=%" PetscInt_FMT " (%" PetscInt_FMT "x%" PetscInt_FMT " grid)\n\n",N,n,m);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+          Compute the problem matrices A and B
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&A);CHKERRQ(ierr);
+  ierr = MatSetSizes(A,PETSC_DECIDE,PETSC_DECIDE,N,N);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(A);CHKERRQ(ierr);
+  ierr = MatSetUp(A);CHKERRQ(ierr);
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&B);CHKERRQ(ierr);
+  ierr = MatSetSizes(B,PETSC_DECIDE,PETSC_DECIDE,N,N);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(B);CHKERRQ(ierr);
+  ierr = MatSetUp(B);CHKERRQ(ierr);
+
+  ierr = MatGetOwnershipRange(A,&Istart,&Iend);CHKERRQ(ierr);
+  for (II=Istart;II<Iend;II++) {
+    i = II/n; j = II-i*n;
+    if (i>0) { ierr = MatSetValue(A,II,II-n,-0.2,INSERT_VALUES);CHKERRQ(ierr); }
+    if (i<m-1) { ierr = MatSetValue(A,II,II+n,-0.2,INSERT_VALUES);CHKERRQ(ierr); }
+    if (j>0) { ierr = MatSetValue(A,II,II-1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    if (j<n-1) { ierr = MatSetValue(A,II,II+1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    ierr = MatSetValue(A,II,II,7.0,INSERT_VALUES);CHKERRQ(ierr);
+    ierr = MatSetValue(B,II,II,2.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  if (Istart==0) {
+    ierr = MatSetValue(B,0,0,6.0,INSERT_VALUES);CHKERRQ(ierr);
+    ierr = MatSetValue(B,0,1,-1.0,INSERT_VALUES);CHKERRQ(ierr);
+    ierr = MatSetValue(B,1,0,-1.0,INSERT_VALUES);CHKERRQ(ierr);
+    ierr = MatSetValue(B,1,1,1.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  ierr = MatAssemblyBegin(A,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(A,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyBegin(B,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(B,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+          Compute sparser approximations A0 and B0
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&A0);CHKERRQ(ierr);
+  ierr = MatSetSizes(A0,PETSC_DECIDE,PETSC_DECIDE,N,N);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(A0);CHKERRQ(ierr);
+  ierr = MatSetUp(A0);CHKERRQ(ierr);
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&B0);CHKERRQ(ierr);
+  ierr = MatSetSizes(B0,PETSC_DECIDE,PETSC_DECIDE,N,N);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(B0);CHKERRQ(ierr);
+  ierr = MatSetUp(B0);CHKERRQ(ierr);
+
+  ierr = MatGetOwnershipRange(A0,&Istart,&Iend);CHKERRQ(ierr);
+  for (II=Istart;II<Iend;II++) {
+    i = II/n; j = II-i*n;
+    if (j>0) { ierr = MatSetValue(A0,II,II-1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    if (j<n-1) { ierr = MatSetValue(A0,II,II+1,-3.0,INSERT_VALUES);CHKERRQ(ierr); }
+    ierr = MatSetValue(A0,II,II,7.0,INSERT_VALUES);CHKERRQ(ierr);
+    ierr = MatSetValue(B0,II,II,2.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  if (Istart==0) {
+    ierr = MatSetValue(B0,0,0,6.0,INSERT_VALUES);CHKERRQ(ierr);
+    ierr = MatSetValue(B0,1,1,1.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  ierr = MatAssemblyBegin(A0,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(A0,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyBegin(B0,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(B0,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+                Create the eigensolver and set various options
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = EPSCreate(PETSC_COMM_WORLD,&eps);CHKERRQ(ierr);
+  ierr = EPSSetOperators(eps,A,B);CHKERRQ(ierr);
+  ierr = EPSSetProblemType(eps,EPS_GHEP);CHKERRQ(ierr);
+  ierr = EPSGetST(eps,&st);CHKERRQ(ierr);
+  ierr = STSetType(st,STSINVERT);CHKERRQ(ierr);
+  mats[0] = A0; mats[1] = B0;
+  ierr = STSetSplitPreconditioner(st,2,mats,SUBSET_NONZERO_PATTERN);CHKERRQ(ierr);
+  ierr = EPSSetTarget(eps,0.0);CHKERRQ(ierr);
+  ierr = EPSSetWhichEigenpairs(eps,EPS_TARGET_MAGNITUDE);CHKERRQ(ierr);
+  ierr = EPSSetFromOptions(eps);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+                 Solve the eigensystem and display solution
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = EPSSolve(eps);CHKERRQ(ierr);
+
+  /* show detailed info unless -terse option is given by user */
+  ierr = PetscOptionsHasName(NULL,NULL,"-terse",&terse);CHKERRQ(ierr);
+  if (terse) {
+    ierr = EPSErrorView(eps,EPS_ERROR_RELATIVE,NULL);CHKERRQ(ierr);
+  } else {
+    ierr = PetscViewerPushFormat(PETSC_VIEWER_STDOUT_WORLD,PETSC_VIEWER_ASCII_INFO_DETAIL);CHKERRQ(ierr);
+    ierr = EPSConvergedReasonView(eps,PETSC_VIEWER_STDOUT_WORLD);CHKERRQ(ierr);
+    ierr = EPSErrorView(eps,EPS_ERROR_RELATIVE,PETSC_VIEWER_STDOUT_WORLD);CHKERRQ(ierr);
+    ierr = PetscViewerPopFormat(PETSC_VIEWER_STDOUT_WORLD);CHKERRQ(ierr);
+  }
+  ierr = EPSDestroy(&eps);CHKERRQ(ierr);
+  ierr = MatDestroy(&A);CHKERRQ(ierr);
+  ierr = MatDestroy(&B);CHKERRQ(ierr);
+  ierr = MatDestroy(&A0);CHKERRQ(ierr);
+  ierr = MatDestroy(&B0);CHKERRQ(ierr);
+  ierr = SlepcFinalize();
+  return ierr;
+}
+
+/*TEST
+
+   testset:
+      args: -eps_nev 4 -terse
+      output_file: output/ex49_1.out
+      requires: !single
+      test:
+         suffix: 1
+      test:
+         suffix: 1_jd
+         args: -eps_type jd -st_type precond
+      test:
+         suffix: 1_lobpcg
+         args: -eps_type lobpcg -st_type precond -eps_smallest_real -st_shift 0.2
+
+   testset:
+      args: -eps_type ciss -eps_all -rg_type ellipse -rg_ellipse_center 0 -rg_ellipse_radius 0.34 -rg_ellipse_vscale .2 -terse
+      output_file: output/ex49_2.out
+      test:
+         suffix: 2
+      test:
+         suffix: 2_nost
+         args: -eps_ciss_usest 0
+         requires: !single
+      test:
+         suffix: 2_par
+         nsize: 2
+         args: -eps_ciss_partitions 2
+         requires: !single
+
+TEST*/