+++ src/sys/classes/st/tests/test7.c
@@ -1,0 +1,134 @@
+/*
+   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+   SLEPc - Scalable Library for Eigenvalue Problem Computations
+   Copyright (c) 2002-2021, Universitat Politecnica de Valencia, Spain
+
+   This file is part of SLEPc.
+   SLEPc is distributed under a 2-clause BSD license (see LICENSE).
+   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+*/
+
+static char help[] = "Test ST with one matrix and split preconditioner.\n\n";
+
+#include <slepcst.h>
+
+int main(int argc,char **argv)
+{
+  Mat            A,Pa,Pmat,mat[1];
+  ST             st;
+  KSP            ksp;
+  PC             pc;
+  Vec            v,w;
+  STType         type;
+  PetscBool      flg;
+  PetscScalar    sigma;
+  PetscInt       n=10,i,Istart,Iend;
+  PetscErrorCode ierr;
+
+  ierr = SlepcInitialize(&argc,&argv,(char*)0,help);if (ierr) return ierr;
+  ierr = PetscOptionsGetInt(NULL,NULL,"-n",&n,NULL);CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"\n1-D Laplacian, n=%" PetscInt_FMT "\n\n",n);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+     Compute the operator matrix for the 1-D Laplacian
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&A);CHKERRQ(ierr);
+  ierr = MatSetSizes(A,PETSC_DECIDE,PETSC_DECIDE,n,n);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(A);CHKERRQ(ierr);
+  ierr = MatSetUp(A);CHKERRQ(ierr);
+
+  ierr = MatGetOwnershipRange(A,&Istart,&Iend);CHKERRQ(ierr);
+  for (i=Istart;i<Iend;i++) {
+    if (i>0) { ierr = MatSetValue(A,i,i-1,-1.0,INSERT_VALUES);CHKERRQ(ierr); }
+    if (i<n-1) { ierr = MatSetValue(A,i,i+1,-1.0,INSERT_VALUES);CHKERRQ(ierr); }
+    ierr = MatSetValue(A,i,i,2.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  ierr = MatAssemblyBegin(A,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(A,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatCreateVecs(A,&v,&w);CHKERRQ(ierr);
+  ierr = VecSet(v,1.0);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+     Compute the split preconditioner matrix (one diagonal)
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&Pa);CHKERRQ(ierr);
+  ierr = MatSetSizes(Pa,PETSC_DECIDE,PETSC_DECIDE,n,n);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(Pa);CHKERRQ(ierr);
+  ierr = MatSetUp(Pa);CHKERRQ(ierr);
+
+  ierr = MatGetOwnershipRange(Pa,&Istart,&Iend);CHKERRQ(ierr);
+  for (i=Istart;i<Iend;i++) {
+    ierr = MatSetValue(Pa,i,i,2.0,INSERT_VALUES);CHKERRQ(ierr);
+  }
+  ierr = MatAssemblyBegin(Pa,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(Pa,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+                Create the spectral transformation object
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = STCreate(PETSC_COMM_WORLD,&st);CHKERRQ(ierr);
+  mat[0] = A;
+  ierr = STSetMatrices(st,1,mat);CHKERRQ(ierr);
+  mat[0] = Pa;
+  ierr = STSetSplitPreconditioner(st,1,mat,SAME_NONZERO_PATTERN);CHKERRQ(ierr);
+  ierr = STSetTransform(st,PETSC_TRUE);CHKERRQ(ierr);
+  ierr = STSetFromOptions(st);CHKERRQ(ierr);
+  ierr = STCayleySetAntishift(st,-0.3);CHKERRQ(ierr);   /* only relevant for cayley */
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+               Form the preconditioner matrix and print it
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = PetscObjectTypeCompareAny((PetscObject)st,&flg,STSINVERT,STCAYLEY,"");CHKERRQ(ierr);
+  if (flg) {
+    ierr = STGetKSP(st,&ksp);CHKERRQ(ierr);
+    ierr = KSPGetPC(ksp,&pc);CHKERRQ(ierr);
+    ierr = STGetOperator(st,NULL);CHKERRQ(ierr);
+    ierr = PCGetOperators(pc,NULL,&Pmat);CHKERRQ(ierr);
+    ierr = MatView(Pmat,NULL);CHKERRQ(ierr);
+  }
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+                    Apply the operator
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  /* sigma=0.0 */
+  ierr = STSetUp(st);CHKERRQ(ierr);
+  ierr = STGetType(st,&type);CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"ST type %s\n",type);CHKERRQ(ierr);
+  ierr = STApply(st,v,w);CHKERRQ(ierr);
+  ierr = VecView(w,NULL);CHKERRQ(ierr);
+
+  /* sigma=0.1 */
+  sigma = 0.1;
+  ierr = STSetShift(st,sigma);CHKERRQ(ierr);
+  ierr = STGetShift(st,&sigma);CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"With shift=%g\n",(double)PetscRealPart(sigma));CHKERRQ(ierr);
+  if (flg) {
+    ierr = STGetOperator(st,NULL);CHKERRQ(ierr);
+    ierr = PCGetOperators(pc,NULL,&Pmat);CHKERRQ(ierr);
+    ierr = MatView(Pmat,NULL);CHKERRQ(ierr);
+  }
+  ierr = STApply(st,v,w);CHKERRQ(ierr);
+  ierr = VecView(w,NULL);CHKERRQ(ierr);
+
+  ierr = STDestroy(&st);CHKERRQ(ierr);
+  ierr = MatDestroy(&A);CHKERRQ(ierr);
+  ierr = MatDestroy(&Pa);CHKERRQ(ierr);
+  ierr = VecDestroy(&v);CHKERRQ(ierr);
+  ierr = VecDestroy(&w);CHKERRQ(ierr);
+  ierr = SlepcFinalize();
+  return ierr;
+}
+
+/*TEST
+
+   test:
+      suffix: 1
+      args: -st_type {{cayley shift sinvert}separate output}
+      requires: !single
+
+TEST*/