+++ src/sys/classes/st/tests/test9.c
@@ -8,15 +8,16 @@
    - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 */
 
-static char help[] = "Test ST with four matrices.\n\n";
+static char help[] = "Test ST with four matrices and split preconditioner.\n\n";
 
 #include <slepcst.h>
 
 int main(int argc,char **argv)
 {
-  Mat            A,B,C,D,mat[4];
+  Mat            A,B,C,D,Pa,Pb,Pc,Pd,Pmat,mat[4];
   ST             st;
   KSP            ksp;
+  PC             pc;
   Vec            v,w;
   STType         type;
   PetscScalar    sigma;
@@ -84,6 +85,50 @@ int main(int argc,char **argv)
   ierr = VecSet(v,1.0);CHKERRQ(ierr);
 
   /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+     Compute the split preconditioner matrices (four diagonals)
+     - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&Pa);CHKERRQ(ierr);
+  ierr = MatSetSizes(Pa,PETSC_DECIDE,PETSC_DECIDE,n,n);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(Pa);CHKERRQ(ierr);
+  ierr = MatSetUp(Pa);CHKERRQ(ierr);
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&Pb);CHKERRQ(ierr);
+  ierr = MatSetSizes(Pb,PETSC_DECIDE,PETSC_DECIDE,n,n);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(Pb);CHKERRQ(ierr);
+  ierr = MatSetUp(Pb);CHKERRQ(ierr);
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&Pc);CHKERRQ(ierr);
+  ierr = MatSetSizes(Pc,PETSC_DECIDE,PETSC_DECIDE,n,n);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(Pc);CHKERRQ(ierr);
+  ierr = MatSetUp(Pc);CHKERRQ(ierr);
+
+  ierr = MatCreate(PETSC_COMM_WORLD,&Pd);CHKERRQ(ierr);
+  ierr = MatSetSizes(Pd,PETSC_DECIDE,PETSC_DECIDE,n,n);CHKERRQ(ierr);
+  ierr = MatSetFromOptions(Pd);CHKERRQ(ierr);
+  ierr = MatSetUp(Pd);CHKERRQ(ierr);
+
+  ierr = MatGetOwnershipRange(Pa,&Istart,&Iend);CHKERRQ(ierr);
+  for (i=Istart;i<Iend;i++) {
+    ierr = MatSetValue(Pa,i,i,2.0,INSERT_VALUES);CHKERRQ(ierr);
+    if (i>0) {
+      ierr = MatSetValue(Pb,i,i,(PetscScalar)i,INSERT_VALUES);CHKERRQ(ierr);
+    } else {
+      ierr = MatSetValue(Pb,i,i,-1.0,INSERT_VALUES);CHKERRQ(ierr);
+    }
+    ierr = MatSetValue(Pd,i,i,i*.1,INSERT_VALUES);CHKERRQ(ierr);
+  }
+
+  ierr = MatAssemblyBegin(Pa,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(Pa,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyBegin(Pb,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(Pb,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyBegin(Pc,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(Pc,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyBegin(Pd,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+  ierr = MatAssemblyEnd(Pd,MAT_FINAL_ASSEMBLY);CHKERRQ(ierr);
+
+  /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
                 Create the spectral transformation object
      - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
@@ -93,69 +138,39 @@ int main(int argc,char **argv)
   mat[2] = C;
   mat[3] = D;
   ierr = STSetMatrices(st,4,mat);CHKERRQ(ierr);
+  mat[0] = Pa;
+  mat[1] = Pb;
+  mat[2] = Pc;
+  mat[3] = Pd;
+  ierr = STSetSplitPreconditioner(st,4,mat,SUBSET_NONZERO_PATTERN);CHKERRQ(ierr);
   ierr = STGetKSP(st,&ksp);CHKERRQ(ierr);
   ierr = KSPSetTolerances(ksp,100*PETSC_MACHINE_EPSILON,PETSC_DEFAULT,PETSC_DEFAULT,PETSC_DEFAULT);CHKERRQ(ierr);
+  ierr = STSetTransform(st,PETSC_TRUE);CHKERRQ(ierr);
   ierr = STSetFromOptions(st);CHKERRQ(ierr);
+  ierr = STGetKSP(st,&ksp);CHKERRQ(ierr);
+  ierr = KSPGetPC(ksp,&pc);CHKERRQ(ierr);
 
   /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-              Apply the transformed operator for several ST's
+                   Apply the operator
      - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
-  /* shift, sigma=0.0 */
+  /* sigma=0.0 */
   ierr = STSetUp(st);CHKERRQ(ierr);
   ierr = STGetType(st,&type);CHKERRQ(ierr);
   ierr = PetscPrintf(PETSC_COMM_WORLD,"ST type %s\n",type);CHKERRQ(ierr);
-  for (i=0;i<4;i++) {
-    ierr = STMatMult(st,i,v,w);
-    ierr = PetscPrintf(PETSC_COMM_WORLD,"k= %" PetscInt_FMT "\n",i);CHKERRQ(ierr);
-    ierr = VecView(w,NULL);CHKERRQ(ierr);
-  }
+  ierr = PCGetOperators(pc,NULL,&Pmat);CHKERRQ(ierr);
+  ierr = MatView(Pmat,NULL);CHKERRQ(ierr);
   ierr = STMatSolve(st,v,w);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"solve\n");CHKERRQ(ierr);
   ierr = VecView(w,NULL);CHKERRQ(ierr);
 
-  /* shift, sigma=0.1 */
+  /* sigma=0.1 */
   sigma = 0.1;
   ierr = STSetShift(st,sigma);CHKERRQ(ierr);
   ierr = STGetShift(st,&sigma);CHKERRQ(ierr);
   ierr = PetscPrintf(PETSC_COMM_WORLD,"With shift=%g\n",(double)PetscRealPart(sigma));CHKERRQ(ierr);
-  for (i=0;i<4;i++) {
-    ierr = STMatMult(st,i,v,w);
-    ierr = PetscPrintf(PETSC_COMM_WORLD,"k= %" PetscInt_FMT "\n",i);CHKERRQ(ierr);
-    ierr = VecView(w,NULL);CHKERRQ(ierr);
-  }
+  ierr = PCGetOperators(pc,NULL,&Pmat);CHKERRQ(ierr);
+  ierr = MatView(Pmat,NULL);CHKERRQ(ierr);
   ierr = STMatSolve(st,v,w);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"solve\n");CHKERRQ(ierr);
-  ierr = VecView(w,NULL);CHKERRQ(ierr);
-
-  /* sinvert, sigma=0.1 */
-  ierr = STPostSolve(st);CHKERRQ(ierr);
-  ierr = STSetType(st,STSINVERT);CHKERRQ(ierr);
-  ierr = STGetType(st,&type);CHKERRQ(ierr);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"ST type %s\n",type);CHKERRQ(ierr);
-  ierr = STGetShift(st,&sigma);CHKERRQ(ierr);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"With shift=%g\n",(double)PetscRealPart(sigma));CHKERRQ(ierr);
-  for (i=0;i<4;i++) {
-    ierr = STMatMult(st,i,v,w);
-    ierr = PetscPrintf(PETSC_COMM_WORLD,"k= %" PetscInt_FMT "\n",i);CHKERRQ(ierr);
-    ierr = VecView(w,NULL);CHKERRQ(ierr);
-  }
-  ierr = STMatSolve(st,v,w);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"solve\n");CHKERRQ(ierr);
-  ierr = VecView(w,NULL);CHKERRQ(ierr);
-
-  /* sinvert, sigma=-0.5 */
-  sigma = -0.5;
-  ierr = STSetShift(st,sigma);CHKERRQ(ierr);
-  ierr = STGetShift(st,&sigma);CHKERRQ(ierr);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"With shift=%g\n",(double)PetscRealPart(sigma));CHKERRQ(ierr);
-  for (i=0;i<4;i++) {
-    ierr = STMatMult(st,i,v,w);
-    ierr = PetscPrintf(PETSC_COMM_WORLD,"k= %" PetscInt_FMT "\n",i);CHKERRQ(ierr);
-    ierr = VecView(w,NULL);CHKERRQ(ierr);
-  }
-  ierr = STMatSolve(st,v,w);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"solve\n");CHKERRQ(ierr);
   ierr = VecView(w,NULL);CHKERRQ(ierr);
 
   ierr = STDestroy(&st);CHKERRQ(ierr);
@@ -163,6 +178,10 @@ int main(int argc,char **argv)
   ierr = MatDestroy(&B);CHKERRQ(ierr);
   ierr = MatDestroy(&C);CHKERRQ(ierr);
   ierr = MatDestroy(&D);CHKERRQ(ierr);
+  ierr = MatDestroy(&Pa);CHKERRQ(ierr);
+  ierr = MatDestroy(&Pb);CHKERRQ(ierr);
+  ierr = MatDestroy(&Pc);CHKERRQ(ierr);
+  ierr = MatDestroy(&Pd);CHKERRQ(ierr);
   ierr = VecDestroy(&v);CHKERRQ(ierr);
   ierr = VecDestroy(&w);CHKERRQ(ierr);
   ierr = SlepcFinalize();
@@ -173,13 +192,7 @@ int main(int argc,char **argv)
 
    test:
       suffix: 1
-      args: -st_transform -st_matmode {{copy shell}}
-      output_file: output/test4_1.out
+      args: -st_type {{shift sinvert}separate output} -st_pc_type jacobi
       requires: !single
-
-   test:
-      suffix: 2
-      args: -st_matmode {{copy shell}}
-      output_file: output/test4_2.out
 
 TEST*/